% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.cor.R
\name{as.cor}
\alias{as.cor}
\alias{as.assoc}
\alias{as.pvalue}
\title{as.XXX conversion functions}
\usage{
as.cor(x)

as.assoc(x)

as.pvalue(x)
}
\arguments{
\item{x}{numeric matrix: matrix to convert}
}
\value{
a matrix with an appropriate class
}
\description{
as.XXX conversion functions
}
