% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.matrix.R
\name{plot.matrix}
\alias{plot.matrix}
\alias{plot}
\title{plot.matrix}
\usage{
\method{plot}{matrix}(x, y = NULL, digits = NA, col = NULL,
  breaks = NULL, key = list(cex.axis = 1), na.col = "white",
  fmt = NULL, fmt.key = fmt, ...)
}
\arguments{
\item{x}{matrix}

\item{y}{unused}

\item{digits}{number of digits for numeric data or length of string for non-numeric data}

\item{col}{a vector of colors or a function, e.g. \code{\link[grDevices]{heat.colors}} with one parameter \code{n}}

\item{breaks}{breaks for numeric values or values for \code{col}}

\item{key}{list of parameters used for \code{\link[graphics]{axis}}. If set to \code{NULL} then no information will be plotted.}

\item{na.col}{color for missing value (default: white)}

\item{fmt}{format string for writring matrix entries, overwrites \code{digits}}

\item{fmt.key}{format string for writring key entries, overwrites \code{digits}. Default ist \code{fmt}}

\item{...}{further parameter given to the \code{\link[graphics]{plot}} command}
}
\value{
a plot
}
\description{
Visualizes a matrix with a heatmap and distinguishes between numeric and non-numeric matrices. 
You may need to modify \code{mar} with the \code{\link[graphics]{par}} command from its default \code{c(5.1,4.1,4.1,2.1)}.
}
\note{
The use of \code{fmt} or \code{fmt.key} have the same restrictions as the use of \code{fmt} in \code{\link[base]{sprintf}}: 

\emph{The format string is passed down the OS's sprintf function, and incorrect formats can cause the latter to crash the R process. 
R does perform sanity checks on the format, but not all possible user errors on all platforms have been tested, and some might 
be terminal.}
}
\examples{
par(mar=c(5.1, 4.1, 4.1, 4.1))
# numeric matrix
x <- matrix(runif(50), nrow=10)
plot(x)
plot(x, key=NULL)
plot(x, key=list(cex.axis=0.5, tick=FALSE))
plot(x, digits=3)
plot(x, breaks=c(0,1), digits=3, cex=0.6)
# logical matrix
m <- matrix(runif(50)<0.5, nrow=10)
plot(m)
plot(m, col=c("red", "blue"))
plot(m, key=NULL, digits=1)
# character matrix
s <- matrix(sample(letters[1:10], 50, replace=TRUE), nrow=10)
plot(s)
plot(s, col=topo.colors)
plot(s, digits=10)
plot(s, digits=10, col=heat.colors(5), breaks=letters[1:5])
# contingency table
tab <- table(round(rnorm(100)), round(rnorm(100)))
plot(unclass(tab))
}
