% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Titanic.cramer}
\alias{Titanic.cramer}
\title{Survival of passengers on the Titanic}
\format{A 4x4 matrix with Cramer's V computed on
  \describe{
    \item{\code{Class}}{1st, 2nd, 3rd, Crew}
    \item{\code{Sex}}{Male, Female}
    \item{\code{Age}}{Child, Adult}
    \item{\code{Survived}}{No, Yes}
}}
\source{
The data are derived from the \link[datasets:Titanic]{Survival of passengers on the Titanic} data set.
}
\usage{
data(Titanic.cramer)
}
\description{
Matrix of Cramer's V computed on the variables economic status (class), sex, age and survival of the 
fate of passengers on the fatal maiden voyage of the ocean liner 'Titanic'.
}
\examples{
data(Titanic.cramer)
plot(as.assoc(Titanic.cramer))
}
\keyword{datasets}
